FS-4446: Create test to check output tracker columns
For the OUTPUT_TRACKER report, the test checks the column order in a report of unscored applications against the column order in a report of scored applications. The scored report will have more columns so we're comparing all the columns in the unscored report against the first x columns in scored report.